Texture memory affine transformations are now available through cupyx.scipy.affine_transform keyword
