+++ cupyx/__init__.py
@@ -10,7 +10,6 @@ from cupyx import time  # NOQA
 from cupyx import scipy  # NOQA
 from cupyx import optimizing  # NOQA
 from cupyx import lapack  # NOQA
-from cupyx import texture  # NOQA
 
 from cupyx._ufunc_config import errstate  # NOQA
 from cupyx._ufunc_config import geterr  # NOQA
