+++ cupyx/scipy/ndimage/interpolation.py
@@ -5,6 +5,7 @@ import cupy
 import numpy
 
 from cupy._core import internal
+from cupyx import _texture
 from cupyx.scipy.ndimage import _util
 from cupyx.scipy.ndimage import _interp_kernels
 from cupyx.scipy.ndimage import _spline_prefilter_core
@@ -309,7 +310,8 @@ def map_coordinates(input, coordinates, output=None, order=3,
 
 
 def affine_transform(input, matrix, offset=0.0, output_shape=None, output=None,
-                     order=3, mode='constant', cval=0.0, prefilter=True):
+                     order=3, mode='constant', cval=0.0, prefilter=True,
+                     texture_memory=False):
     """Apply an affine transformation.
 
     Given an output image pixel index vector ``o``, the pixel value is
@@ -352,6 +354,10 @@ def affine_transform(input, matrix, offset=0.0, output_shape=None, output=None,
             0.0
         prefilter (bool): It is not used yet. It just exists for compatibility
             with :mod:`scipy.ndimage`.
+        texture_memory (bool): If True, uses GPU texture memory. Supports only
+            2D and 3D single-channel arrays; modes ``'constant'`` and
+            ``'nearest'``; order 0 (nearest neighbor) and 1 (bi/tri-linear
+            interpolation).
 
     Returns:
         cupy.ndarray or None:
@@ -360,6 +366,11 @@ def affine_transform(input, matrix, offset=0.0, output_shape=None, output=None,
 
     .. seealso:: :func:`scipy.ndimage.affine_transform`
     """
+
+    if texture_memory:
+        tm_interp = 'linear' if order > 0 else 'nearest'
+        return _texture.affine_transformation(input, matrix, tm_interp, mode,
+                                              cval)
 
     _check_parameter('affine_transform', order, mode)
