--- cupyx/texture/__init__.py
@@ -1,1 +1,0 @@
-from cupyx.texture._affine import affine_transformation  # NOQA
